+++ astropy/modeling/functional_models.py
@@ -136,17 +136,18 @@ class Gaussian2D(Parametric2DModel):
         Mean of the Gaussian in y.
     x_stddev : float
         Standard deviation of the Gaussian in x.
-        x_stddev and y_stddev must be specified unless a covariance
-        matrix (cov_matrix) is input.
+        ``x_stddev`` and ``y_stddev`` must be specified unless a covariance
+        matrix (``cov_matrix``) is input.
     y_stddev : float
         Standard deviation of the Gaussian in y.
-        x_stddev and y_stddev must be specified unless a covariance
-        matrix (cov_matrix) is input.
+        ``x_stddev`` and ``y_stddev`` must be specified unless a covariance
+        matrix (``cov_matrix``) is input.
     theta : float, optional
-        Rotation angle in radians. The rotation angle increases clockwise.
+        Rotation angle in radians. The rotation angle increases
+        counterclockwise.
     cov_matrix : ndarray, optional
-        A 2x2 covariance matrix. If specified, overrides the x_stddev,
-        y_stddev, and theta specification.
+        A 2x2 covariance matrix. If specified, overrides the ``x_stddev``,
+        ``y_stddev``, and ``theta`` specification.
 
     Notes
     -----
@@ -163,16 +164,39 @@ class Gaussian2D(Parametric2DModel):
             a = \\left(\\frac{\\cos^{2}{\\left (\\theta \\right )}}{2 \\sigma_{x}^{2}} +
             \\frac{\\sin^{2}{\\left (\\theta \\right )}}{2 \\sigma_{y}^{2}}\\right)
 
-            b = \\left(\\frac{-\\sin{\\left (2 \\theta \\right )}}{2 \\sigma_{x}^{2}} +
+            b = \\left(\\frac{\\sin{\\left (2 \\theta \\right )}}{2 \\sigma_{x}^{2}} -
             \\frac{\\sin{\\left (2 \\theta \\right )}}{2 \\sigma_{y}^{2}}\\right)
 
             c = \\left(\\frac{\\sin^{2}{\\left (\\theta \\right )}}{2 \\sigma_{x}^{2}} +
             \\frac{\\cos^{2}{\\left (\\theta \\right )}}{2 \\sigma_{y}^{2}}\\right)
 
+    If using a ``cov_matrix``, the model is of the form:
+        .. math::
+            f(x, y) = A e^{-0.5 \\left(\\vec{x} - \\vec{x}_{0}\\right)^{T} \\Sigma^{-1} \\left(\\vec{x} - \\vec{x}_{0}\\right)}
+
+    where :math:`\\vec{x} = [x, y]`, :math:`\\vec{x}_{0} = [x_{0}, y_{0}]`,
+    and :math:`\\Sigma` is the covariance matrix:
+
+        .. math::
+            \\Sigma = \\left(\\begin{array}{ccc}
+            \\sigma_x^2               & \\rho \\sigma_x \\sigma_y \\\\
+            \\rho \\sigma_x \\sigma_y & \\sigma_y^2
+            \end{array}\\right)
+
+    :math:`\\rho` is the correlation between `x` and `y`, which should
+    be between -1 and +1.  Positive correlation corresponds to a
+    ``theta`` in the range 0 to 90 degrees.  Negative correlation
+    corresponds to a ``theta`` in the range of 0 to -90 degrees.
+
+    See [1]_ for more details about the 2D Gaussian function.
 
     See Also
     --------
     Gaussian1D, Box2D, Beta2D
+
+    References
+    ----------
+    .. [1] http://en.wikipedia.org/wiki/Gaussian_function
     """
 
     amplitude = Parameter()
@@ -222,7 +246,7 @@ class Gaussian2D(Parametric2DModel):
         xdiff = x - x_mean
         ydiff = y - y_mean
         a = 0.5 * ((cost2 / xstd2) + (sint2 / ystd2))
-        b = 0.5 * (-(sin2t / xstd2) + (sin2t / ystd2))
+        b = 0.5 * ((sin2t / xstd2) - (sin2t / ystd2))
         c = 0.5 * ((sint2 / xstd2) + (cost2 / ystd2))
         return amplitude * np.exp(-((a * xdiff ** 2) + (b * xdiff * ydiff) +
                                     (c * ydiff ** 2)))
@@ -246,16 +270,16 @@ class Gaussian2D(Parametric2DModel):
         xdiff2 = xdiff ** 2
         ydiff2 = ydiff ** 2
         a = 0.5 * ((cost2 / xstd2) + (sint2 / ystd2))
-        b = 0.5 * (-(sin2t / xstd2) + (sin2t / ystd2))
+        b = 0.5 * ((sin2t / xstd2) - (sin2t / ystd2))
         c = 0.5 * ((sint2 / xstd2) + (cost2 / ystd2))
         g = amplitude * np.exp(-((a * xdiff2) + (b * xdiff * ydiff) +
                                  (c * ydiff2)))
         da_dtheta = (sint * cost * ((1. / ystd2) - (1. / xstd2)))
         da_dx_stddev = -cost2 / xstd3
         da_dy_stddev = -sint2 / ystd3
-        db_dtheta = (-cos2t / xstd2) + (cos2t / ystd2)
-        db_dx_stddev = sin2t / xstd3
-        db_dy_stddev = -sin2t / ystd3
+        db_dtheta = (cos2t / xstd2) - (cos2t / ystd2)
+        db_dx_stddev = -sin2t / xstd3
+        db_dy_stddev = sin2t / ystd3
         dc_dtheta = -da_dtheta
         dc_dx_stddev = -sint2 / xstd3
         dc_dy_stddev = -cost2 / ystd3